Fix -Wformat warning

+++ include/sandbox/utils/logger.h
@@ -153,7 +153,7 @@ private:
 
         int bytes_written = snprintf(format_buffer, sizeof(format_buffer),
                                      "\033\[%dm%s%s\033\[0m",
-                                     color, prefix, msg);
+                                     (int)color, prefix, msg);
         sbAssert(bytes_written >= 0
                      && bytes_written < (int)sizeof(format_buffer),
                  "format_buffer too small");
